SSCS-2688 [AS] Refactor before implementing subscriptions integration tests
